Created simple format
You can add separate css and js file in respective folder and aldd that file in head section in order.
And if wnting to add new html file then add below the endex.html file and add link in index.html.
So it help to identify each section or changes or portion of each part.